datatypes made uniform

+++ algorithm/brute_force/brute_force.h
@@ -6,14 +6,14 @@ class BruteForce
 {
 private:
     int m, n;   // number of candidate sites, number of clients
-    vector<int> construction;
-    vector<vector<int>> cost;
+    vector<long double> construction;
+    vector<vector<long double>> cost;
     vector<long double> traffic;
 
     vector<int> allotment;
 
 public:
-    BruteForce(int _m, int _n, vector<int> &_construction, vector<vector<int>> &_cost, vector<long double> &_traffic):
+    BruteForce(int _m, int _n, vector<long double> &_construction, vector<vector<long double>> &_cost, vector<long double> &_traffic):
         m(_m), n(_n), construction(_construction), cost(_cost), traffic(_traffic), allotment(vector<int>(n, -1))
     {}
